Adding support for path based SVGs
The current version of the library only supports SVGs which have been constructed from shapes. SVGs can also be constructed from paths. If the path attributes are attached to the `svg` element, this data is currently stripped from the sprite, leading to icons which look filled in. This change allows the library to support both filled and path based SVGs.